move 'prop' tests to separate file
