+++ types/ramda/tsconfig.json
@@ -180,6 +180,7 @@
         "test/prepend-tests.ts",
         "test/product-tests.ts",
         "test/project-tests.ts",
+        "test/prop-tests.ts",
         "test/propEq-tests.ts"
     ]
 }
